requirements.txt
djangorestframework
django==4.2
black
djangorestframework
djangorestframework-simplejwt